backend/requirements.txt
Flask==2.3.2
flask-cors==3.1.0
opencv-python-headless==4.8.0.74
numpy==1.24.3
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==9.4.0
openpyxl==3.1.2
python-dotenv==1.0.0
uuid==1.30
